# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: E-Logbook Enterprise API
  description: |
    API for the E-Logbook Enterprise platform — a medical resident case-logging
    and faculty evaluation system with multi-tenant support, SSO, AI-powered
    educational insights, payment integration, and SCIM user provisioning.

    ## Authentication
    Most endpoints require JWT authentication via Supabase Auth. Pass the access
    token as a Bearer token in the `Authorization` header. Public endpoints
    (health, SSO check) do not require authentication.

    Edge Functions use Supabase's built-in JWT verification through the
    shared `authenticate()` helper.
  version: "1.0.0"
  contact:
    name: E-Logbook Support
    url: https://elogbook.dev
servers:
  - url: https://app.elogbook.dev
    description: Production server
  - url: http://localhost:3000
    description: Local development
  - url: https://{tenant}.elogbook.dev
    description: Tenant-specific subdomain
    variables:
      tenant:
        default: demo
        description: Tenant slug

# ============================================================================
# Components
# ============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT access token
    ServiceRole:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase service_role key (admin operations)
...
paths:
...
  /api/{tenant}/reports/specialty.csv:
    parameters:
      - name: tenant
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: exportSpecialtyCsv
      summary: Export specialty distribution as CSV
      description: |
        Exports case distribution by specialty as a CSV file.
        Supports optional date filtering.
      tags: [Reports]
      security:
        - BearerAuth: []
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: CSV file download
          content:
            text/csv:
              schema:
                type: string
        "400":
          description: Query error
        "401":
          description: Unauthorized
        "403":
          description: Invalid tenant

  # ============= Reports: Status CSV =============
  /api/{tenant}/reports/status.csv:
    parameters:
      - name: tenant
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: exportStatusCsv
      summary: Export case status distribution as CSV
      description: |
        Exports case counts by status (draft, pending, approved, rejected)
        as a CSV file. Supports optional date filtering.
      tags: [Reports]
      security:
        - BearerAuth: []
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: CSV file download
          content:
            text/csv:
              schema:
                type: string
        "400":
          description: Query error
        "401":
          description: Unauthorized
        "403":
          description: Invalid tenant

# ============================================================================
# Paths — Supabase Edge Functions
# ============================================================================
...